240425_03
레벨업 추가, 버그 수정

+++ Dungeon.cs
@@ -9,9 +9,9 @@ namespace SpartaDungeonGame
     internal class Dungeon
     {
         public string dungeonName;
-        public int requiredAmr;
+        public float requiredAmr;
 
-        public Dungeon(string dungeonName, int requiredAmr)
+        public Dungeon(string dungeonName, float requiredAmr)
         {
             this.dungeonName = dungeonName;
             this.requiredAmr = requiredAmr;
